requirements.txt
fastapi
uvicorn
openai
google-generativeai
requests
python-dotenv